NU-1091 rewrite Errors class component to function component
